% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCS_class.R
\name{BCS}
\alias{BCS}
\alias{dBCS}
\alias{pBCS}
\alias{qBCS}
\alias{rBCS}
\title{The Box-Cox Symmetric Class of Distributions}
\usage{
dBCS(x, mu, sigma, lambda, nu = NULL, gen = "NO")

pBCS(q, mu, sigma, lambda, nu = NULL, gen = "NO")

qBCS(p, mu, sigma, lambda, nu = NULL, gen = "NO")

rBCS(n, mu, sigma, lambda, nu = NULL, gen = "NO", rep = 1L)
}
\arguments{
\item{x, q}{vector of non-negative quantiles.}

\item{mu}{vector of location parameter values.}

\item{sigma}{vector of dispersion parameter values.}

\item{lambda}{vector of skewness parameter values.}

\item{nu}{vector of right tail/kurtosis parameter values. Not all
distributions are indexed by this parameter. In this case,
\code{nu = NULL}.}

\item{gen}{character specification passed as argument to \code{\link{BCSgen}}. It
specifies the generating distribution for the BCS class of distributions.
A table with the current available generating distributions can be seen
in details.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}

\item{rep}{number of replicates with size \code{n} to return.}
}
\value{
\code{dBCS} returns the density function, \code{pBCS}
gives the distribution function, \code{qBCS} gives the quantile function,
and \code{rBCS} generates random observations.
}
\description{
Probability mass function, cumulative distribution function,
quantile function, and random generation for the Box-Cox symmetric (BCS)
class of distributions.
}
\details{
This set of functions represents the probability function, the
    cumulative distribution function, quantile function, and a random number
    generator for the Box-Cox symmetric class of distributions proposed by
    Ferrari and Fumes (2017). This class of distributions includes the
    Box-Cox t (Rigby and Stasinopoulos, 2006), Box-Cox Cole-Green
    (or Box-Cox normal; Cole and Green, 1992), Box-Cox power exponential
    (Rigby and Stasinopoulos, 2004) distributions, and the class of the
    log-symmetric distributions (Vanegas and Paula, 2016) as special cases.
    The current available generating distributions can be seen below.
  \tabular{lll}{
 \bold{Distribution}  \tab \bold{Abbreviation} \tab \bold{Does it have an extra parameter?}\cr
 Cauchy  \tab \code{"CA"}      \tab  no  \cr
 Canonical slash  \tab \code{"CSL"}      \tab  no  \cr
 Double exponential (Laplace)  \tab \code{"DE"}      \tab  no  \cr
 Logistic  \tab \code{"LO"}      \tab  no  \cr
 Normal  \tab \code{"NO"}      \tab  no  \cr
 Power exponential  \tab \code{"PE"}      \tab  yes  \cr
 Slash  \tab \code{"SL"}      \tab  yes  \cr
 Student-t  \tab \code{"ST"}      \tab  yes  \cr
 }
}
\examples{
\dontrun{
## Box-Cox t
y <- rBCS(500, 8, 0.15, 1, 10, gen = "ST")

par(mfrow = c(1, 2))
hist(y, prob = TRUE, main = "Box Cox t")
curve(dBCS(x, 8, 0.15, 1, 10, gen = "ST"), add = TRUE, col = 2, lwd = 2)
plot(ecdf(y), main = "", xlab = "y")
curve(pBCS(x, 8, 0.15, 1, 10, gen = "ST"), add = TRUE, col = 2, lwd = 2)

## Box-Cox Cole-Green
y <- rBCS(500, 8, 0.15, 1, 10, gen = "NO")

hist(y, prob = TRUE, main = "Box Cox Cole-Green")
curve(dBCS(x, 8, 0.15, 1, 10, gen = "NO"), add = TRUE, col = 2, lwd = 2)
plot(ecdf(y), main = "", xlab = "y")
curve(pBCS(x, 8, 0.15, 1, 10, gen = "NO"), add = TRUE, col = 2, lwd = 2)

## Box-Cox Cauchy
y <- rBCS(500, 8, 0.15, -1.2, gen = "CA")

hist(y, prob = TRUE, main = "Box-Cox Cauchy")
curve(dBCS(x, 8, 0.15, -1.2, gen = "CA"), add = TRUE, col = 2, lwd = 2)
plot(ecdf(y), main = "", xlab = "y")
curve(pBCS(x, 8, 0.15, -1.2, gen = "CA"), add = TRUE, col = 2, lwd = 2)

## Box-Cox power exponential
y <- rBCS(500, 8, 0.15, 1, 10, gen = "PE")

hist(y, prob = TRUE, main = "Box Cox PE")
curve(dBCS(x, 8, 0.15, 1, 10, gen = "PE"), add = TRUE, col = 2, lwd = 2)
plot(ecdf(y), main = "")
curve(pBCS(x, 8, 0.15, 1, 10, gen = "PE"), add = TRUE, col = 2, lwd = 2)
}
}
\references{
Cole, T., & Green, P.J. (1992). Smoothing reference centile curves: the LMS
     method and penalized likelihood. Stat. Med, 11, 1305--1319.

 Ferrari, S. L., & Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. AStA Advances in Statistical Analysis,
     101, 321--344.

 Rigby, R. A., & Stasinopoulos, D. M. (2004). Smooth centile curves for skew
     and kurtotic data modelled using the Box-Cox power exponential
     distribution. Statistics in medicine, 23, 3053--3076.

 Rigby, R. A., & Stasinopoulos, D. M. (2006). Using the Box-Cox t
     distribution in GAMLSS to model skewness and kurtosis. Statistical
     Modelling, 6, 209-229.

 Vanegas, L. H., & Paula, G. A. (2016). Log-symmetric distributions:
     statistical properties and parameter estimation. Brazilian Journal of
     Probability and Statistics, 30, 196--220.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
