% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_functions.R
\name{BCSgen}
\alias{BCSgen}
\title{Box-Cox Symmetric Distribution Class Generators}
\usage{
BCSgen(dist)
}
\arguments{
\item{dist}{Character specification of the generating distribution, see
details.}
}
\value{
The above family of distribution returns a list whose components are
    set of functions referring to the generating function of the Box-Cox
    symmetric class of distributions. The output components of each function
    are:
 \itemize{
   \item{d:}{ Density generating function \code{function(u, nu)}.}
   \item{p:}{ Cumulative distribution function \code{function(q, nu)}.}
   \item{q:}{ Quantile function \code{function(p, nu)}.}
   \item{extraP:}{ Logical; it receives \code{TRUE} if the generating
     distribution has an extra parameter.}
   \item{weigh:}{ Weighting function \code{function(z, nu)}.}
 }

 The arguments of the returned functions are:
 \describe{
   \item{\code{u, q}}{ Vector of positive quantiles.}
   \item{\code{p}}{ Vector of probabilities.}
   \item{\code{z}}{ A vector of transformed responses with the generalized
    Box-Cox transformation.}
   \item{\code{nu}}{ Extra parameter of the generating distribution, if it
    does not exist, it receives \code{NULL}.}
 }
}
\description{
Current available generating distributions that can be used to
    fit a model belonging to the Box-Cox symmetric class of distributions.
}
\details{
There are some distributions available for the density generating
    function. The following table display their names and their abbreviations
    to be passed to \code{BCSgen()}.

  \tabular{lll}{
 \bold{Distribution}  \tab \bold{Abbreviation} \tab \bold{It has an extra parameter?}\cr
 Cauchy  \tab \code{"CA"}      \tab  no  \cr
 Canonical slash  \tab \code{"CSL"}      \tab  no  \cr
 Double exponential (Laplace)  \tab \code{"DE"}      \tab  no  \cr
 Logistic  \tab \code{"LO"}      \tab  no  \cr
 Normal  \tab \code{"NO"}      \tab  no  \cr
 Power exponential  \tab \code{"PE"}      \tab  yes  \cr
 Slash  \tab \code{"SL"}      \tab  yes  \cr
 Student-t  \tab \code{"ST"}      \tab  yes  \cr
 }
}
\examples{
\dontrun{
curve(BCSgen("NO")$d(x^2), xlim = c(-5, 5),
      ylim = c(0, 0.7), ylab = "Density")
curve(BCSgen("CA")$d(x^2), add = TRUE, col = 2)
curve(BCSgen("CSL")$d(x^2), add = TRUE, col = 3)
curve(BCSgen("DE")$d(x^2), add = TRUE, col = 4)
curve(BCSgen("LO")$d(x^2), add = TRUE, col = 6)
legend("topright", c("Cauchy", "Canonical Slash",
                     "Laplace", "Logistic", "Normal"),
        col = c(2:4, 6, 1), lty = 1, bty = "n")
}

}
\references{
Ferrari, S. L., & Fumes, G. (2017). Box–Cox symmetric
 distributions and applications to nutritional data. AStA Advances in
 Statistical Analysis, 101(3), 321-344.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
