% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_correlations.R
\name{cormat.mbcsec}
\alias{cormat.mbcsec}
\alias{ind}
\alias{un}
\alias{arma}
\alias{cluster}
\alias{matern}
\title{Working Correlation Matrices for Multivariate Box-Cox Symmetric
    Generated by Elliptical Copula Distributions}
\usage{
ind(d)

un(d)

arma(d, p = 0, q = 0)

cluster(
  d,
  id,
  type = c("independence", "ar1", "ma1", "exchangeable", "unstructured")
)

matern(d, D, smt = 0.5)
}
\arguments{
\item{d}{Dimension of the correlation matrix.}

\item{p, q}{Order of the autoregressive and the moving average component,
respectively, passed as arguments to \code{arma()}.}

\item{id}{Subject id for longitudinal/clustered data. This is a vector of
the same lenght of the number of observations. Please note that data
must be sorted in way that observations from the same cluster are
contiguous.}

\item{type}{A character string specifying the correlation structure among
groups for longitudinal/clustered data. At the
moment, the following are implemented:
\tabular{ll}{
\code{independence}  \tab Working independence. \cr
\code{ar1}  \tab Autoregressive of order 1. \cr
\code{ma1}  \tab Moving average of order 1. \cr
\code{exchangeable1}  \tab Exchangeable. \cr
\code{unstructured}  \tab Unstructured \cr
}}

\item{D}{Matrix with values of the distances between pairs of data
locations for spatial data.}

\item{smt}{Value of the shape parameter of the Matern correlation class.
The default \code{smt = 0.5} corresponds to an exponential correlation
model.}
}
\description{
Current available correlation matrices in the \code{mbcsec}
    package.
}
\details{
The functions related to the association structures are a direct
    adaptation of the functions of the \code{\link[gcmr]{gcmr}} package.
    The documentation of the original functions of the package can be seen
    at \code{\link[gcmr]{cormat.gcmr}} documentation. The available
    correlation matrices are:

 \tabular{ll}{
 \bold{Function}  \tab \bold{Correlation}\cr
 \code{ind}  \tab Working independence. \cr
 \code{un}  \tab Unstructured \cr
 \code{arma}  \tab ARMA(p, q) \cr
 \code{cluster}  \tab Longitudinal/clustered data \cr
 \code{matern}  \tab Matern spatial correlation \cr
 }

 Each type of structure will require different arguments for the association
 matrix P to be completed. However, all functions return a list of the
 following components:

 \itemize{
 \item{\code{npar}:}{ Number of parameters associated with the correlation
     structure.}
 \item{\code{start}:}{ A function \code{function(y)} that returns the
     initial value for use in optimization algorithms. Its argument is the
     vector/matrix of observations \code{y}.}
 \item{\code{P}:}{ A function \code{function(alpha)} which returns the
     association matrix itself. Its argument is the parameter vector
     associated with the correlation matrix structure.}
 \item{\code{name}:}{ Name of the specified correlation structure.}
 }
}
\references{
Guido Masarotto, & Cristiano Varin (2017). Gaussian Copula
    Regression in R. Journal of Statistical Software, 77(8), 1--26.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
