% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBCSec_class.R
\name{MBCSec}
\alias{MBCSec}
\alias{dmbcsec}
\alias{rmbcsec}
\alias{print.mbcsec.data}
\alias{plot.mbcsec.data}
\title{Multivariate Box-Cox Symmetric Class of Distributions Generating
    by Elliptical Copula}
\usage{
dmbcsec(
  x,
  param,
  P = NULL,
  df = 4,
  copula = c("gaussian", "t", "cauchy", "dexponential"),
  gen = "NO"
)

rmbcsec(
  n,
  param,
  P = NULL,
  d = 2L,
  df = NULL,
  copula = c("gaussian", "t", "cauchy", "dexponential"),
  gen = "NO"
)

\method{print}{mbcsec.data}(x, ...)

\method{plot}{mbcsec.data}(x, method = c("pearson", "kendall", "spearman"), ...)
}
\arguments{
\item{x}{Vector or matrix of non-negative quantiles. If \code{x} is a matrix,
each row is taken to be a quantile. For methods, \code{x} is
a \code{"mbcsec.data"} object.}

\item{param}{A list that has as components the vectors of marginal
parameters \code{mu}, \code{sigma}, \code{lambda}, and \code{nu}. In case
the marginal distribution is not indexed by \code{nu}, it receives
\code{NULL} by default.}

\item{P}{Association matrix. If it is specified with \code{NULL} (default),
then the identity matrix is used.}

\item{df}{Degrees of freedom associated with the t copula.}

\item{copula}{Character; informs which distribution in the class of
elliptical distributions should be used to generate the elliptical
copula. Currently, the copulas available are:
Gaussian (\code{"gaussian"}), Student's t (\code{"t"}), Cauchy
(\code{"cauchy"}), and double exponential (\code{"dexponential"}).}

\item{gen}{A character or a vector character; specifies the distributions
generating the marginal BCS distributions. If all BCS margins are
generated by the same generating density, it is sufficient to enter only
one character. A table with the current available generating
distributions for the BCS class can be seen in \code{\link{BCSgen}}.}

\item{n}{Number of random values to return.}

\item{d}{Dimension of the data to be generated. Argument used only in
\code{rmbcsec()}.}

\item{...}{Further arguments for other specific methods.}

\item{method}{A character string passed to the plot method indicating which
correlation coefficient is to be computed. One of \code{"pearson"}
(default), \code{"kendall"}, or \code{"spearman"}: can be abbreviated.}
}
\value{
\code{dmbcsec} returns the evaluated joint density function and
 \code{rmbcsec} returns a \code{"mbcsec.data"} object that contains the
     observations generated and the specifications for the multivariate
     distribution. More specifically, it returns a list with the
     following components:

 \itemize{
   \item{data:}{ Random observations generated from the multivariate
     distribution specified via \code{rmbcsec()}.}
   \item{spec:}{ A list with the distribution specifications:
     marginal parameters (\code{param}), association matrix (\code{P}),
     degrees of freedom (\code{df}), copula (\code{copula}), and
     generating functions for margins (\code{gen}).}
 }
}
\description{
These functions provide the joint probability density function
    and a random generator for the multivariate Box-Cox symmetric class of
    distributions generated by elliptical copulas.
}
\examples{
### Sample size and dimension
n <- 500
d <- 3

### Association matrix
P <- matrix(c(1, 0.7, -0.5, 0.7, 1, 0.2, -0.5, 0.2, 1), 3, 3)

### Marginals specifications

# Marginals
gen <- c("NO", "ST", "DE")

# Marginal parameters
param <- list(mu = c(10, 6, 1), sigma = c(0.15, 0.15, 0.1),
              lambda = c(-2, 0, 2), nu = c(NA, 2, NA))

### Copula
copula <- "gaussian"

### Generating observations
y <- rmbcsec(n, param, P, d, copula = copula, gen = gen)

y
plot(y)


}
\references{
Ferrari, S. L., & Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. AStA Advances in Statistical Analysis,
     101, 321--344.

 Vanegas, L. H., & Paula, G. A. (2016). Log-symmetric distributions:
     statistical properties and parameter estimation. Brazilian Journal of
     Probability and Statistics, 30, 196--220.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
