% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBCSec_class.R
\name{MBCSec}
\alias{MBCSec}
\alias{dmbcsec}
\alias{rmbcsec}
\title{Multivariate Box-Cox Symmetric Class of Distributions Generating
    by Elliptical Copula}
\usage{
dmbcsec(
  x,
  param,
  P = NULL,
  df = 4,
  copula = c("gaussian", "st", "cauchy", "dexponential"),
  gen = "NO"
)

rmbcsec(
  n,
  param,
  P = NULL,
  d = 2L,
  df = 4,
  copula = c("gaussian", "st", "cauchy", "dexponential"),
  gen = "NO"
)
}
\arguments{
\item{x}{vector or matrix of non-negative quantiles. If \code{x} is a matrix,
each row is taken to be a quantile.}

\item{param}{a list that has as components the vectors of marginal
parameters \code{mu}, \code{sigma}, \code{lambda}, and \code{nu}. In case
the marginal distribution is not indexed by \code{nu}, it receives
\code{NULL} by default.}

\item{P}{association matrix. If it is specified with \code{NULL} (default),
then the identity matrix is used.}

\item{df}{degrees of freedom associated with the t copula.}

\item{copula}{character; informs which distribution in the class of
elliptical distributions should be used to generate the elliptical
copula. Currently, the copulas available are:
Gaussian (\code{"gaussian"}), Student's t (\code{"st"}), Cauchy
(\code{"cauchy"}), and double exponential (\code{"dexponential"}).}

\item{gen}{a character or a vector character; specifies the distributions
generating the marginal BCS distributions. If all BCS margins are
generated by the same generating density, it is sufficient to enter only
one character. A table with the current available generating
distributions for the BCS class can be seen in \code{\link{BCSgen}}.}

\item{n}{number of random values to return.}

\item{d}{dimension of the data to be generated. Argument used only in
\code{rmbcsec()}.}
}
\value{
\code{dmbcsec} returns the joint density function and \code{rmbcsec}
    generates random observations.
}
\description{
These functions provide the joint probability density function
    and a random generator for the multivariate Box-Cox symmetric class of
    distributions generated by elliptical copulas.
}
\references{
Cole, T., & Green, P.J. (1992). Smoothing reference centile curves: the LMS
     method and penalized likelihood. Stat. Med, 11, 1305--1319.

 Ferrari, S. L., & Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. AStA Advances in Statistical Analysis,
     101, 321--344.

 Rigby, R. A., & Stasinopoulos, D. M. (2004). Smooth centile curves for skew
     and kurtotic data modelled using the Box-Cox power exponential
     distribution. Statistics in medicine, 23, 3053--3076.

 Rigby, R. A., & Stasinopoulos, D. M. (2006). Using the Box-Cox t
     distribution in GAMLSS to model skewness and kurtosis. Statistical
     Modelling, 6, 209-229.

 Vanegas, L. H., & Paula, G. A. (2016). Log-symmetric distributions:
     statistical properties and parameter estimation. Brazilian Journal of
     Probability and Statistics, 30, 196--220.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
